Fix wrong processing of the "Blue Armor Class" string from a DEH
Description:
P_GiveArmor(player, armortype) is inlined in doom2.exe. There are two entries of GiveArmor(*,blue_armor_class) in the sources.
The first is for MegaArmor and the second is for Supercharge. If you change the "Blue Armor Class" value in a DEH, then dehacked replaces only the first entry in the EXE (MegaArmor entrie). So for the second (Supercharge entrie) we must always use armortype=2 instead of blue_armor_class value from a DEH