Count par time silently in background
Only bump the state when the par time has finished (silently) counting
up - for demo compatibility reasons we cannot bump the state early even
if no gunshots are being played[1]

[1] If the user clicks while the par time is still counting up, another
click is required to enter the next level. This needs to happen even if
par time is suppressed - if the demo was recorded with two clicks, and
playback suddenly only needs one, an extra gunshot at the start of the
next map will cause a desync. (Similar to the issue described in the
previous commit; could be fixed in a new compatibility level)