configure.ac: use AM_SILENT_RULES if available
Rationale: it's easier to see warnings during the build process as
they're not lost within a storm of three-foot-long compiler commands.

Check the macro exists before trying to use it though, as it is new
to Automake 1.11.

"./configure --disable-silent-rules" or "make V=1" will reactivate
verbose rules, in case you still need to see the entire command lines
used for compilation.